EKF: add output observer option
Add a parameter selectable option to output a more kinematically position and velocity. This will have degraded tracking relative to the EKF states, but the velocity will be closer to the first derivative of the position and reduce the effect inertial prediction errors on control loops that are operating in a pure velocity feedback mode.